+++ examples/historical_metadata_get_dataset_condition.py
@@ -7,6 +7,6 @@ if __name__ == "__main__":
     key = "YOUR_API_KEY"
     client = Historical(key=key)
 
-    conditions = client.metadata.list_dataset_conditions(dataset="XNAS.ITCH")
+    condition = client.metadata.get_dataset_condition(dataset="XNAS.ITCH")
 
-    pprint(conditions)
+    pprint(condition)